pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lmgamerl"
version = "0.1.0"
description = "Multi-Turn RL Training System with AgentTrainer for Language Model Game Reinforcement Learning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "lmgamerl Team"},
]
keywords = [
    "reinforcement learning", 
    "PPO", 
    "multi-turn", 
    "training", 
    "AI", 
    "machine learning",
    "language models",
    "game environments"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # CORE DEPENDENCIES - ORDER MATTERS
    #"torch>=2.7.0",
    #"flash-attn>=2.8.0.post2",
    
    # MACHINE LEARNING & DEEP LEARNING FRAMEWORKS
    "transformers==4.52.4",
    "accelerate",
    "peft",
    "datasets",
    "safetensors>=0.4.3",
    "huggingface_hub>=0.21.0",
    "tokenizers",
    
    # REINFORCEMENT LEARNING & DISTRIBUTED COMPUTING
    "ray[default]>=2.10",
    "tensordict==0.6.2",
    "torchdata",
    "vllm>=0.8.5",
    
    # VERL FRAMEWORK DEPENDENCIES
    "liger-kernel",
    "dill",
    "codetiming",
    "hydra-core",
    "pylatexenc",
    "pre-commit",
    
    # ENVIRONMENT & AGENT DEPENDENCIES
    "gym_sokoban",
    "gymnasium[toy-text]",
    "wandb",
    
    # DATA PROCESSING & UTILITIES
    "pandas",
    "numpy",
    "pyarrow>=19.0.0",
    "pybind11",
    "packaging>=20.0",
    "tqdm",
    "scipy",
    "plotly",
    "matplotlib",
    "matplotlib-inline",
    
    # DEVELOPMENT & DEBUGGING
    "IPython",
    "debugpy",
    "jupyter",
    "notebook",
    
    # WEB INTERFACE & API
    "fastapi",
    "uvicorn",
    "pydantic",
    
    # SYSTEM & UTILITIES
    "psutil",
    "pyyaml",
    "filelock",
    "typing-extensions",
    "sympy",
    "jinja2",
    "networkx",
    "setuptools",
    
    # WEBSHOP & DATASET DEPENDENCIES (included by default)
    "gdown>=4.6.6",
    "spacy>=3.4.0",
    
    # JAVA & NLP (for WebShop)
    # Note: Java JDK and Maven are installed via conda/apt in setup
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0", 
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]
jupyter = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0", 
    "ipywidgets>=8.0.0",
]

[project.scripts]
lmgamerl-train = "lmgamerl.train:main"

[project.urls]
Homepage = "https://github.com/mignonjia/lmgame_train_reorg"
"Bug Reports" = "https://github.com/mignonjia/lmgame_train_reorg/issues"
"Source" = "https://github.com/mignonjia/lmgame_train_reorg"

[tool.setuptools.packages.find]
where = ["."]
include = ["lmgamerl*"]
exclude = ["tests*", "scripts*", "verl*", "external*"]

[tool.setuptools.package-data]
"lmgamerl" = ["*.md", "*.txt", "*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 120